Move uplift from util to own namespace
